+++ R/rcpp.R
@@ -1,3 +1,0 @@
-#' @useDynLib nngeo
-#' @importFrom Rcpp sourceCpp
-NULL