+++ R/st_nn_pnt_geo.R
@@ -1,14 +1,11 @@
 .st_nn_pnt_geo = function(x, y, k, maxdist, progress) {
 
   x_features = length(sf::st_geometry(x))
+  y_features = length(sf::st_geometry(y))
   x_coord = sf::st_coordinates(x)
   y_coord = sf::st_coordinates(y)
   maxdist = units::set_units(maxdist, "m")
   maxdist = as.numeric(maxdist)
-
-  # result = getIdsDists(x_coord, y_coord, k, maxdist)
-  # ids = result[[1]]
-  # dist_matrix = result[[2]]
 
   ids = matrix(NA, nrow = x_features, ncol = k)
   dist_matrix = matrix(NA, nrow = x_features, ncol = k)
@@ -20,7 +17,14 @@
 
   for(i in 1:x_features) {
 
-    dists = getDistancesN(x_coord[i, ], y_coord)
+    dists = rep(NA, y_features)
+
+    for(j in 1:y_features) {
+
+      dists[j] = .Call(addr_dist_one, c(x_coord[i, ], y_coord[j, ]))
+
+    }
+
     ids1 = order(dists)[1:k]
     ids[i, ] = ids1
     dist_matrix[i, ] = dists[ids1]